+++ src/main/java/org/patternfly/component/actionlist/ActionList.java
@@ -29,7 +29,8 @@ import static org.patternfly.layout.Classes.modifier;
 /**
  * An action list is a group of actions, controls, or buttons with set spacing.
  *
- * @see <a href="https://www.patternfly.org/components/action-list/html">https://www.patternfly.org/components/action-list/html</a>
+ * @see <a href=
+ *      "https://www.patternfly.org/components/action-list/html">https://www.patternfly.org/components/action-list/html</a>
  */
 public class ActionList extends BaseComponent<HTMLElement, ActionList> {