+++ src/main/java/org/patternfly/component/expandable/ExpandableSection.java
@@ -44,6 +44,7 @@ import static org.patternfly.layout.Classes.expanded;
 import static org.patternfly.layout.Classes.limitWidth;
 import static org.patternfly.layout.Classes.modifier;
 import static org.patternfly.layout.Size.lg;
+import static org.patternfly.layout.Variable.componentVar;
 
 /**
  * An expandable section component is used to support progressive disclosure in a form or page by hiding additional content when
@@ -98,7 +99,7 @@ public class ExpandableSection extends BaseComponent<HTMLDivElement, ExpandableS
                 toggle.removeIcon();
             }
             if (truncate != DEFAULT_TRUNCATE && content != null) {
-                content.style("--pf-v5-c-expandable-section--m-truncate__content--LineClamp: " + truncate);
+                componentVar(component(expandableSection), "m-truncate__content", "LineClamp").style(content, truncate);
             }
         }
         if (detachedFromId != null) {