+++ src/main/java/org/patternfly/layout/VariableScope.java
@@ -13,20 +13,11 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.patternfly.component;
+package org.patternfly.layout;
 
-import elemental2.dom.MutationRecord;
+public enum VariableScope {
 
-/**
- * Interface meant to be implemented by subcomponents who need a reference to the base component.
- * <p>
- * The first call to {@link #passComponent(BaseComponent)} is typically initiated by the base component in the
- * {@link org.jboss.elemento.Attachable#attach(MutationRecord)} method (assuming that the base component implements
- * {@link org.jboss.elemento.Attachable}).
- */
-public interface ComponentReference<C extends BaseComponent<?, ?>> {
+    global,
 
-    void passComponent(C component);
-
-    C mainComponent();
+    component
 }